requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncmy
weaviate-client
langchain
langgraph
pydantic-settings
google-auth
python-jose[cryptography]
aiofiles
langchain-google-genai
python-dotenv
google-cloud-secret-manager

# Dev tools
ruff
black
